+++ desktop/cmp/grid/find/impl/GridFindFieldImplModel.ts
@@ -186,72 +186,7 @@ export class GridFindFieldImplModel extends HoistModel {
     }
 
     private getRecords(): StoreRecord[] {
-        if (!this._records) {
-            const records = this.sortRecordsRecursive([...this.gridModel.store.rootRecords]);
-            this._records = this.sortRecordsByGroupBy(records);
-        }
-        return this._records;
-    }
-
-    // Sort records with GridModel's sortBy(s) using the Column's comparator
-    private sortRecordsRecursive(records: StoreRecord[]): StoreRecord[] {
-        const {gridModel} = this,
-            {sortBy, treeMode, agApi, store} = gridModel,
-            ret: StoreRecord[] = [];
-
-        [...sortBy].reverse().forEach(it => {
-            const column = gridModel.getColumn(it.colId);
-            if (!column) return;
-
-            const {field, getValueFn} = column,
-                compFn = (column.getAgSpec().comparator as Function).bind(column),
-                direction = it.sort === 'desc' ? -1 : 1;
-
-            const ctx = {field, column, gridModel, store, agParams: null};
-            records.sort((a, b) => {
-                const valueA = getValueFn({record: a, ...ctx}),
-                    valueB = getValueFn({record: b, ...ctx}),
-                    nodeA = agApi?.getRowNode(a.agId),
-                    nodeB = agApi?.getRowNode(b.agId);
-
-                return compFn(valueA, valueB, nodeA, nodeB) * direction;
-            });
-        });
-
-        records.forEach(rec => {
-            ret.push(rec);
-            if (treeMode && !isEmpty(rec.children)) {
-                const children = this.sortRecordsRecursive(rec.children);
-                ret.push(...children);
-            }
-        });
-
-        return ret;
-    }
-
-    // Sort records with GridModel's groupBy(s) using the GridModel's groupSortFn
-    private sortRecordsByGroupBy(records: StoreRecord[]) {
-        const {gridModel} = this,
-            {agApi, groupBy, groupSortFn, store} = gridModel;
-
-        [...groupBy].reverse().forEach(groupField => {
-            const column = gridModel.getColumn(groupField);
-            if (!column) return;
-
-            const {field, getValueFn} = column,
-                ctx = {field, column, gridModel, store, agParams: null};
-
-            records.sort((a, b) => {
-                const valueA = getValueFn({record: a, ...ctx}),
-                    valueB = getValueFn({record: b, ...ctx}),
-                    nodeA = agApi?.getRowNode(a.agId),
-                    nodeB = agApi?.getRowNode(b.agId);
-
-                return groupSortFn(valueA, valueB, field, {gridModel, nodeA, nodeB});
-            });
-        });
-
-        return records;
+        return (this._records ??= this.gridModel.getSortedRecords());
     }
 
     private getActiveFields(): string[] {